Remove functions that reported success they had not established
groth16Verify and plonkVerify checked the protocol string and the public signal count and then returned true. A verifier that returns true without performing the pairing check accepts every proof, including a forged one. This library has no pairing implementation to build a real check from: no Fp2/Fp6/Fp12, no G2 group law, no Miller loop, no final exponentiation, in TypeScript or in either addon. The provers were equally hollow, emitting proofs whose elements were zero. All six entry points now live in snarkjs/unimplemented.ts, throw, and are deliberately not re-exported. msmGPU returned the identity point with usedGPU: true; nttGPU returned its input unchanged with the same flag. Both now throw, and the former behaviour is recorded in comments so it is not reintroduced. sme.ts reported usedSME from isAvailable() while running a plain JS loop; it now reports the fallback that actually ran. hardware-report fabricated cpuPercent+30 and gpuPercent+70. Removed. radix4.ts delegated every function to the radix-2 implementation, which made ntt-consistency.prop.test.ts -- asserting radix-2 equals radix-4 -- tautological. It is now a real radix-4 butterfly. hardware_detect.cc inferred AMX from the CPU brand string. AMX is not detectable from user space, so the capability is removed rather than guessed; SME now queries hw.optional.arm.FEAT_SME, which exists. isAMXAccelerated returns a three-state value, and the test asserting it was a boolean was itself asserting the removed claim. Packaging: the addon filename searched for did not match the prebuild produced, and native-rust/target was shipped. Both corrected.

+++ native/include/metal_gpu.h
@@ -52,7 +52,9 @@ typedef struct {
     bool device_available;
     bool unified_memory;
     int max_threads_per_group;
-    int max_buffer_length;
+    /* Bytes. Must not be `int`: Apple Silicon reports values above 2^31
+     * (e.g. ~3.1 GB on an M4 Max), which truncated to a negative number. */
+    uint64_t max_buffer_length;
     char device_name[256];
 } MetalGPUStatus;
